# openapi.yaml
openapi: 3.0.0
info:
  title: petsonline API
  description: API REST completa con autenticación JWT y CRUD de entidades
  version: 1.0.0

servers:
  - url: http://localhost:3012/api
    description: Servidor de desarrollo

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: admin@sistema.com
        password:
          type: string
          example: Admin123456

    RegisterRequest:
      type: object
      required:
        - nombre
        - email
        - password
        - role
      properties:
        nombre:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string

    User:
      type: object
      properties:
        _id:
          type: string
        nombre:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string

    EntityBase:
      type: object
      properties:
        _id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        imagen:
          type: string
        imagenThumbnail:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string

    Error:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string

paths:
  /auth/login:
    post:
      tags:
        - Autenticación
      summary: Iniciar sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login exitoso
        '401':
          description: Credenciales inválidas

  /auth/register:
    post:
      tags:
        - Autenticación
      summary: Registrar usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuario registrado
        '409':
          description: Email ya existe

  /auth/profile:
    get:
      tags:
        - Autenticación
      summary: Obtener perfil
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil obtenido

  /auth/users:
    get:
      tags:
        - Autenticación
      summary: Listar usuarios (Admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de usuarios

  /upload/image:
    post:
      tags:
        - Upload
      summary: Subir imagen
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida

  /mascota:
    get:
      tags:
        - Mascota
      summary: Listar todos los Mascotas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Mascotas
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Mascota
      summary: Crear Mascota
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Mascota creado

  /mascota/{id}:
    get:
      tags:
        - Mascota
      summary: Obtener Mascota por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Mascota encontrado
        '404':
          description: Mascota no encontrado
    patch:
      tags:
        - Mascota
      summary: Actualizar Mascota
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Mascota actualizado
    delete:
      tags:
        - Mascota
      summary: Eliminar Mascota
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Mascota eliminado

  /mascota/{id}/upload-image:
    post:
      tags:
        - Mascota
      summary: Subir imagen a Mascota
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida

  /cliente:
    get:
      tags:
        - Cliente
      summary: Listar todos los Clientes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Clientes
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Cliente
      summary: Crear Cliente
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Cliente creado

  /cliente/{id}:
    get:
      tags:
        - Cliente
      summary: Obtener Cliente por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cliente encontrado
        '404':
          description: Cliente no encontrado
    patch:
      tags:
        - Cliente
      summary: Actualizar Cliente
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Cliente actualizado
    delete:
      tags:
        - Cliente
      summary: Eliminar Cliente
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cliente eliminado

  /cliente/{id}/upload-image:
    post:
      tags:
        - Cliente
      summary: Subir imagen a Cliente
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida

  /producto:
    get:
      tags:
        - Producto
      summary: Listar todos los Productos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Productos
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Producto
      summary: Crear Producto
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Producto creado

  /producto/{id}:
    get:
      tags:
        - Producto
      summary: Obtener Producto por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Producto encontrado
        '404':
          description: Producto no encontrado
    patch:
      tags:
        - Producto
      summary: Actualizar Producto
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Producto actualizado
    delete:
      tags:
        - Producto
      summary: Eliminar Producto
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Producto eliminado

  /producto/{id}/upload-image:
    post:
      tags:
        - Producto
      summary: Subir imagen a Producto
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida

  /servicio-veterinario:
    get:
      tags:
        - ServicioVeterinario
      summary: Listar todos los ServicioVeterinarios
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de ServicioVeterinarios
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - ServicioVeterinario
      summary: Crear ServicioVeterinario
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: ServicioVeterinario creado

  /servicio-veterinario/{id}:
    get:
      tags:
        - ServicioVeterinario
      summary: Obtener ServicioVeterinario por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ServicioVeterinario encontrado
        '404':
          description: ServicioVeterinario no encontrado
    patch:
      tags:
        - ServicioVeterinario
      summary: Actualizar ServicioVeterinario
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: ServicioVeterinario actualizado
    delete:
      tags:
        - ServicioVeterinario
      summary: Eliminar ServicioVeterinario
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ServicioVeterinario eliminado

  /servicio-veterinario/{id}/upload-image:
    post:
      tags:
        - ServicioVeterinario
      summary: Subir imagen a ServicioVeterinario
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida

  /reserva:
    get:
      tags:
        - Reserva
      summary: Listar todos los Reservas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Reservas
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Reserva
      summary: Crear Reserva
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Reserva creado

  /reserva/{id}:
    get:
      tags:
        - Reserva
      summary: Obtener Reserva por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reserva encontrado
        '404':
          description: Reserva no encontrado
    patch:
      tags:
        - Reserva
      summary: Actualizar Reserva
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Reserva actualizado
    delete:
      tags:
        - Reserva
      summary: Eliminar Reserva
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reserva eliminado

  /reserva/{id}/upload-image:
    post:
      tags:
        - Reserva
      summary: Subir imagen a Reserva
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida